requirements.txt
bittensor~=9.11.1
ratelimit==2.2.1
cachetools==5.3.1
tabulate>=0.9.0
python-dotenv>=1.0.0
backoff==2.2.1
redis==5.2.1
async-timeout>=4.0.0
aioredis>=2.0.0
torch>=2.0.0
transformers>=4.51.3
ijson>=3.2.3
psutil>=5.9.0
gunicorn>=21.2.0
flask>=2.0.0
flask-cors>=3.0.0
requests>=2.28.0
aiomysql>=0.2.0
mysql-connector-python>=8.0.0

imagehash==4.3.1
huggingface_hub[hf_transfer]>=0.27.0
diffusers[torch]==0.34.0
safetensors
loguru
git+https://github.com/openai/CLIP.git

cryptography>=41.0.0

httpx>=0.24.0
tenacity>=8.2.0

asyncio-mqtt>=0.13.0

pillow==10.4.0
nltk>=3.9.1

pydantic

# Development dependencies
pytest==8.3.5
pytest-cov==6.1.1
responses==0.25.7
ruff==0.11.7
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# Additional dependencies for API server
markovify>=0.9.0
diskcache>=5.6.0
datasets>=2.14.0
#fiber @ git+https://github.com/rayonlabs/fiber.git@2.5.0

#async-substrate-interface>=1.3.1
substrate-interface>=1.7.0
scalecodec>=1.2.6

opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation>=0.41b0
opentelemetry-exporter-otlp>=1.20.0
opentelemetry-instrumentation-flask>=0.41b0
opentelemetry-instrumentation-requests>=0.41b0
opentelemetry-instrumentation-mysql>=0.41b0
opentelemetry-instrumentation-redis>=0.41b0